Feature/collection

+++ metabase/permissions.go
@@ -6,18 +6,18 @@ import (
 )
 
 type Group struct {
-	ID          int    `json:"id"`
+	ID          int64  `json:"id"`
 	Name        string `json:"name"`
-	MemberCount int    `json:"member_count"`
+	MemberCount int64  `json:"member_count"`
 }
 
 type Membership struct {
-	MembershipID int `json:"membership_id"`
-	GroupID      int `json:"group_id"`
-	UserID       int `json:"user_id"`
+	MembershipID int64 `json:"membership_id"`
+	GroupID      int64 `json:"group_id"`
+	UserID       int64 `json:"user_id"`
 }
 
-func (c *Client) DeletePermissionsGroup(groupID int) error {
+func (c *Client) DeletePermissionsGroup(groupID int64) error {
 
 	err := c.deleteRequest(fmt.Sprintf("/api/permissions/group/%d", groupID))
 	if err != nil {
@@ -27,7 +27,7 @@ func (c *Client) DeletePermissionsGroup(groupID int) error {
 	return nil
 }
 
-func (c *Client) DeletePermissionsMembership(membershipID int) error {
+func (c *Client) DeletePermissionsMembership(membershipID int64) error {
 
 	err := c.deleteRequest(fmt.Sprintf("/api/permissions/membership/%d", membershipID))
 	if err != nil {
@@ -53,7 +53,7 @@ func (c *Client) GetPermissionsGroups() ([]Group, error) {
 	return res, nil
 }
 
-func (c *Client) GetPermissionsGroup(groupID int) (*Group, error) {
+func (c *Client) GetPermissionsGroup(groupID int64) (*Group, error) {
 
 	resData, err := c.getRequest(fmt.Sprintf("/api/permissions/group/%d", groupID), nil)
 	if err != nil {
@@ -85,11 +85,11 @@ func (c *Client) GetPermissionsMemberships() (map[string][]Membership, error) {
 	return res, nil
 }
 
-func (c *Client) CreatePermissionsMembership(groupID int, userID int) (*Membership, error) {
+func (c *Client) CreatePermissionsMembership(groupID int64, userID int64) (*Membership, error) {
 
 	d := struct {
-		GroupID int `json:"group_id"`
-		UserID  int `json:"user_id"`
+		GroupID int64 `json:"group_id"`
+		UserID  int64 `json:"user_id"`
 	}{
 		groupID,
 		userID,
@@ -139,7 +139,7 @@ func (c *Client) CreatePermissionsGroup(groupName string) (*Group, error) {
 	return &res, nil
 }
 
-func (c *Client) UpdatePermissionsGroup(groupID int, groupName string) (*Group, error) {
+func (c *Client) UpdatePermissionsGroup(groupID int64, groupName string) (*Group, error) {
 
 	d := struct {
 		Name string `json:"name"`
@@ -170,7 +170,7 @@ func (c *Client) UpdatePermissionsGroup(groupID int, groupName string) (*Group, 
 // ================
 
 type PermissionsGraph struct {
-	Revision int                                 `json:"revision"`
+	Revision int64                               `json:"revision"`
 	Groups   map[string](map[string]interface{}) `json:"groups"`
 }
